200.dockter.1 (#203)
* #200 migrate to C.longlong and remove deprecated methods

* #200 Prepare for versioned release

+++ cmd/root.go
@@ -99,7 +99,7 @@ func RunE(_ *cobra.Command, _ []string) error {
 		Port:                           viper.GetInt(option.GrpcPort.Arg),
 		SenzingEngineConfigurationJson: senzingEngineConfigurationJson,
 		SenzingModuleName:              viper.GetString(option.EngineModuleName.Arg),
-		SenzingVerboseLogging:          viper.GetInt(option.EngineLogLevel.Arg),
+		SenzingVerboseLogging:          viper.GetInt64(option.EngineLogLevel.Arg),
 	}
 	return grpcserver.Serve(ctx)
 }
